Refactor user profile component, change ngOnInit to handle stored user data more securely

+++ src/app/user-profile/user-profile.component.ts
@@ -11,12 +11,18 @@ import { FormsModule } from '@angular/forms';
 export class UserProfileComponent implements OnInit {
   userData: any = {};
 
-  constructor(public fetchApiData: FetchApiDataService, public router: Router) {
-    this.userData = JSON.parse(localStorage.getItem('user') || '');
-  }
+  constructor(
+    public fetchApiData: FetchApiDataService,
+    public router: Router
+  ) {}
 
   ngOnInit(): void {
-    this.getUser();
+    const storedUser = JSON.parse(localStorage.getItem('user') || '');
+    if (storedUser) {
+      this.userData = storedUser;
+      // Fetch latest data from API
+      this.getUser();
+    }
   }
 
   openMovies(): void {
@@ -24,9 +30,8 @@ export class UserProfileComponent implements OnInit {
   }
 
   getUser(): void {
-    this.fetchApiData
-      .getUserData(this.userData.Username)
-      .subscribe((response: any) => {
+    this.fetchApiData.getUserData(this.userData.Username).subscribe(
+      (response: any) => {
         this.userData = {
           ...response,
           id: response._id,
@@ -34,23 +39,36 @@ export class UserProfileComponent implements OnInit {
         };
         console.log(this.userData);
         localStorage.setItem('user', JSON.stringify(this.userData));
-      });
+      },
+      (error: any) => {
+        console.error('Error fetching user data:', error);
+      }
+    );
   }
 
   updateUser(): void {
-    this.fetchApiData.updateUser(this.userData.Username, this.userData).subscribe(
-      (response: any) => {
-        this.userData = {
-          ...response,
-          id: response._id,
-          password: this.userData.Password,
-        };
-        localStorage.setItem('user', JSON.stringify(this.userData));
-        console.log("Update successful");
-      },
-      (error: any) => {
-        console.error(error);
-      }
-    );
+    if (
+      window.confirm('Are you sure you want to update? This cannot be undone.')
+    ) {
+      this.fetchApiData
+        .updateUser(this.userData.Username, this.userData)
+        .subscribe(
+          (response: any) => {
+            // Save the updated user info to local storage
+            this.userData = {
+              ...response,
+              id: response._id,
+              password: '', // Clear password after update for security
+            };
+            localStorage.setItem('user', JSON.stringify(this.userData));
+
+            // Fetch latest user info to reflect changes in the view
+            this.getUser();
+          },
+          (error: any) => {
+            console.error('Error updating user: ', error);
+          }
+        );
+    }
   }
 }
